refactor: begin migration to Qt
The Electron version of this app is unable to maintain support for Widevine DRM (required for streaming services) due to anti-competitive practises by Google who maintain Widevine. The move to Qt is necessary to maintain the application's functionality. This commit is a minimal Qt application which will be developed to reach feature parity with the Electron version then merged and released to replace the current Electron version of the app.